Avoid size_t to avoid warnings.

+++ rts/c/util.h
@@ -135,8 +135,8 @@ static void str_builder(struct str_builder *b, const char *s, ...) {
   b->used += needed;
 }
 
-static int lexical_realloc(char **error, unsigned char **ptr, size_t *old_size, size_t new_size) {
-  unsigned char *new = realloc(*ptr, new_size);
+static int lexical_realloc(char **error, unsigned char **ptr, int64_t *old_size, int64_t new_size) {
+  unsigned char *new = realloc(*ptr, (size_t)new_size);
   if (new == NULL) {
     *error = msgprintf("Failed to allocate memory.\nAttempted allocation: %12lld bytes\n",
                        (long long) new_size);
